refactor to customGear
