+++ addons/customGear/fauxLAMG/CfgWeapons.hpp
